+++ Block/Tab/Content/Log.php
@@ -7,8 +7,6 @@ class Log extends \ADM\QuickDevBar\Block\Tab\Panel
 
     protected $_jsonHelper;
 
-    protected $_qdbHelper;
-
     /**
      * @param \Magento\Framework\View\Element\Template\Context $context
      * @param \Magento\Framework\Json\EncoderInterface $jsonEncoder
@@ -18,13 +16,12 @@ class Log extends \ADM\QuickDevBar\Block\Tab\Panel
         \Magento\Framework\View\Element\Template\Context $context,
         \Magento\Framework\Json\Helper\Data $jsonHelper,
         \ADM\QuickDevBar\Helper\Data $qdbHelper,
+        \ADM\QuickDevBar\Helper\Register $qdbHelperRegister,
         array $data = []
     ) {
         $this->_jsonHelper = $jsonHelper;
 
-        $this->_qdbHelper= $qdbHelper;
-
-        parent::__construct($context, $data);
+        parent::__construct($context, $qdbHelper, $qdbHelperRegister, $data);
     }
 
     public function getTailLines()
@@ -34,12 +31,12 @@ class Log extends \ADM\QuickDevBar\Block\Tab\Panel
 
     public function getLogFiles()
     {
-        return $this->_qdbHelper->getLogFiles();
+        return $this->helper->getLogFiles();
     }
 
     public function getJsonLogFiles()
     {
-        return $this->_jsonHelper->jsonEncode($this->_qdbHelper->getLogFiles());
+        return $this->_jsonHelper->jsonEncode($this->helper->getLogFiles());
     }
 
     public function getUrlLog($action)