+++ Block/Tab/Content/Request.php
@@ -4,25 +4,12 @@ namespace ADM\QuickDevBar\Block\Tab\Content;
 
 class Request extends \ADM\QuickDevBar\Block\Tab\Panel
 {
-    /**
-     * @var \ADM\QuickDevBar\Helper\Register
-     */
-    private $qdbHelperRegister;
+
 
     public function getTitle()
     {
         return 'Request';
     }
-
-    public function __construct(
-        \Magento\Framework\View\Element\Template\Context $context,
-        \ADM\QuickDevBar\Helper\Register $qdbHelperRegister,
-        array $data = []
-    ) {
-        $this->qdbHelperRegister = $qdbHelperRegister;
-        parent::__construct($context, $data);
-    }
-
 
     public function getRequestData()
     {